[Refactor] Separate Member service implementations
