+++ src/main/java/boombimapi/domain/member/application/service/MemberServiceV1.java
@@ -4,17 +4,24 @@ package boombimapi.domain.member.application.service;
 import boombimapi.domain.member.presentation.dto.member.req.MemberLeaveReq;
 import boombimapi.domain.member.presentation.dto.member.res.*;
 
+import boombimapi.domain.member.presentation.dto.member.res.mypage.GetMemberResV1;
+import boombimapi.domain.member.presentation.dto.member.res.mypage.GetMemberResV2;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.util.List;
 
-public interface MemberService {
+public interface MemberServiceV1 {
 
     // 1번 구간
-    GetMemberRes getMember(String userId);
+    GetMemberResV1 getMember(String userId);
 
 
+    // 3번 구간
+    List<MyPageVoteRes> getMyVoteAnswer(String userId);
+
+    // 4번 구간
+    List<MyPageVoteRes> getMyVoteQuestion(String userId);
 
     void updateNickname(String userId, String name);
 
@@ -25,16 +32,6 @@ public interface MemberService {
 
     // 프로필 수정
     ProfileRes updateProfile(String userId, MultipartFile multipartFile) throws IOException;
-
-
-    /**
-     * 폐지
-     */
-    /*// 3번 구간
-    List<MyPageVoteRes> getMyVoteAnswer(String userId);
-
-    // 4번 구간
-    List<MyPageVoteRes> getMyVoteQuestion(String userId);*/
 
 
 }
